requirements.txt
flask
Flask-Session
werkzeug
requests
python-dotenv
pdfkit

numpy
pandas
scikit-learn
scipy
xgboost

sentence-transformers
transformers
torch
spacy
nltk

groq
google-generativeai

PyMuPDF
python-docx